I think it's working
Unfortunately the changes broke a bunch of the SupercellEnumerator
tests, so I had to regenerate the files needed for them.

Also, I'm not convinced the old niggli routines were even creating
niggli cells ever time. The result is not that most of the generated
lattices have different transformation matrices and names.