Splitting tests up, I got this.
Need to continue the process onto all the tests, and update configure.ac
to make the .in files to run tests. The test .in files will be in charge
of going to the right directory to run the tests from, set the PATH so
that it calls the just compiled casm, and run whatever is needed.

For some of the tests, it's probably better to downgrade them to simple
bash scripts, since calling Popen just to run some command line
arguments seems like overkill. For now I'm not going to worry about it
though, and just run everything how it it's designed now.