magnetism in properties.calc.json and casm query
Added back in relaxed_mag and relaxed_magmom both to
properties.calc.json (as relaxed_mag_basis, and appropriate
mods to the Relax routine, and the various io handlers), and
as a queryable command in casm.

You must add "relaxed_magmom" into the properties list in
.casm/project_settings for these options to be utilized; support
is not yet full, nor even really partial.