Slash Lattice::is_supercell_of, rename and reorganize
There is now is_superlattice (still in Lattice.hh), and now there is
is_equivalent_superlattice, which is the same as is_superlattice, but
takes a point group (or whatever). The latter has been moved to
crystallography/SymTools.hh.

General things in Lattice.hh have changed from "cell" to "lattice". Also
superdupercell is now make_superduperlattice, and is next in the queue
to get split into make_equivalent_superduperlattice that lives in
SymTools.hh.