Propagated changes to SymCompare constructors
The constructors take a shared_ptr to the prim. Since currently PrimClex
and functions are passing const references, the construction requires
making a copy of the prim to stuff inside the necessary prim_ptr. The
plan is to eventually change PrimClex to hold a shared_ptr, which can
get directly accessed.