"Stupid" template specialization and traits inhertance hybird
Both ways seem to work now. The stupid method is a *lot* easier to
implement, and doesn't need nearly as much new code. The traits method
is near impossible to comprehend and requires traits classes everywhere.
I dream of a templated SymCompare made up entirely of functors and no
inheritance.